Extract `get-config` tests from `sinon-test`.
