+++ test/sinon-test.js
@@ -8,37 +8,6 @@
 
     buster.testCase("sinon", {
 
-        "Function.prototype.toString": {
-
-        },
-
-        ".getConfig": {
-            "gets copy of default config": function () {
-                var config = sinon.getConfig();
-
-                refute.same(config, sinon.defaultConfig);
-                assert.equals(config.injectIntoThis, sinon.defaultConfig.injectIntoThis);
-                assert.equals(config.injectInto, sinon.defaultConfig.injectInto);
-                assert.equals(config.properties, sinon.defaultConfig.properties);
-                assert.equals(config.useFakeTimers, sinon.defaultConfig.useFakeTimers);
-                assert.equals(config.useFakeServer, sinon.defaultConfig.useFakeServer);
-            },
-
-            "should override specified properties": function () {
-                var config = sinon.getConfig({
-                    properties: ["stub", "mock"],
-                    useFakeServer: false
-                });
-
-                refute.same(config, sinon.defaultConfig);
-                assert.equals(config.injectIntoThis, sinon.defaultConfig.injectIntoThis);
-                assert.equals(config.injectInto, sinon.defaultConfig.injectInto);
-                assert.equals(config.properties, ["stub", "mock"]);
-                assert.equals(config.useFakeTimers, sinon.defaultConfig.useFakeTimers);
-                assert.isFalse(config.useFakeServer);
-            }
-        },
-
         ".log": {
             "does nothing gracefully": function () {
                 refute.exception(function () {
